--- src/dynamic-html-text/index.ts
@@ -1,1 +1,0 @@
-export { DynamicHTMLText } from "./dynamic-html-text";
