+++ src/styled-text/index.ts
@@ -1,0 +1,2 @@
+export { StyledHTMLText } from "./styled-html-text";
+export { StyledText } from "./styled-text";
